payload/ubootcli: add the u-boot cli
This is a first cut at the u-boot cli. It builds and loads and runs
and has no commands. That's next.

Interestingly, this is code right from our u-boot tree, and it's full
of style violations which I don't intend to fix.

Remove dependency on anything else, but this won't work until the
libpayload serial drivers for arm are set up.

BUG=None
TEST=build and boot on nyan and get a u-boot cli prompt.
BRANCH=None

Change-Id: I52cf7c02e2bdd00a82e6e48fd2471416c87e4627
Signed-off-by: Ronald G. Minnich <rminnich@google.com>
Reviewed-on: https://chromium-review.googlesource.com/179721
Reviewed-by: Puneet Kumar <puneetster@chromium.org>
Commit-Queue: Puneet Kumar <puneetster@chromium.org>
Tested-by: Puneet Kumar <puneetster@chromium.org>