tegra124: Change the display hack to use window A
The hack seems to set up coreboot display to use window B.  We eventually
want to use the same window as the kernel is going to use (I think), so
that's what this patch does.

We think window B is hiding the contents of window A, which is why we
weren't seeing ChromeOS UI come up.  This fix makes that not happen anymore
by making coreboot use window A.

BUG=chrome-os-partner:24844
TEST=Can boot ChromeOS to UI from coreboot.

Change-Id: I24b95200ba2e8eaeadecd45392ccee5e270aa7da
Reviewed-on: https://chromium-review.googlesource.com/182001
Tested-by: Andrew Chew <achew@nvidia.com>
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Tested-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-by: Jimmy Zhang <jimmzhang@nvidia.com>
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>